split up models into separate files, other cleanup
