+++ src/routes/TelegramWebhookRoute.ts
@@ -1,8 +1,10 @@
 import { Request, Response } from "express";
 import { Op } from "sequelize";
-import { Contact, TelegramID, User } from "../db/db";
+import { Contact } from "../db/models/Contact";
+import { User } from "../db/models/User";
 import { strings_en } from "../strings";
 import { sendTelegramMessage } from "../telegram";
+import { TelegramID } from "../types";
 
 interface TelegramWebhookRequest extends Request {
   body: {
@@ -27,22 +29,22 @@ export function TelegramWebhookRoute(
         "Thanks for using OurSejahtera! Let's stay safer together <3"
       );
     } else {
-    const messageText = req.body.message.text;
-    const telegramID = req.body.message.from.id;
-    if (messageText.toLowerCase() == "/covidpositive") {
-      userInfected(telegramID, (result) => {
-        if (result.saved) {
-          sendTelegramMessage(
-            telegramID,
-            strings_en.telegram_inform_positive,
-          );
-          informContacts(telegramID);
-        } else {
-          sendTelegramMessage(telegramID, "Sorry, something went wrong.");
-        }
-      });
+      const messageText = req.body.message.text;
+      const telegramID = req.body.message.from.id;
+      if (messageText.toLowerCase() == "/covidpositive") {
+        userInfected(telegramID, (result) => {
+          if (result.saved) {
+            sendTelegramMessage(
+              telegramID,
+              strings_en.telegram_inform_positive
+            );
+            informContacts(telegramID);
+          } else {
+            sendTelegramMessage(telegramID, "Sorry, something went wrong.");
+          }
+        });
+      }
     }
-  }
   } catch (e) {
     console.log("Could not get Telegram Message");
   }
@@ -72,7 +74,10 @@ function informContacts(telegramID: TelegramID) {
             },
           }).then((otherPerson) => {
             otherPerson &&
-              sendTelegramMessage(otherPerson.telegram, strings_en.telegram_inform_infect);
+              sendTelegramMessage(
+                otherPerson.telegram,
+                strings_en.telegram_inform_infect
+              );
           });
         });
       });
